[client-api] Move firestore config to a spearate file
